requirements.txt
Adafruit-Blinka==8.6.1
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-mcp3xxx==1.4.12
adafruit-circuitpython-requests==1.12.9
adafruit-circuitpython-typing==1.8.2
Adafruit-PlatformDetect==3.32.0
Adafruit-PureIO==1.1.9
altgraph==0.17.3
astroid==2.12.10
async-timeout==4.0.2
autopep8==1.7.0
blessings==1.7
branca==0.5.0
cachetools==5.2.0
cairocffi==1.4.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.1.3
colorama==0.4.6
colorzero==1.1
colour==0.1.5
commonmark==0.9.1
contourpy==1.0.5
cryptography==38.0.3
cycler==0.11.0
Cython==0.29.32
dbus-python==1.2.16
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.5.1
distro==1.5.0
earthengine-api==0.1.326
ee==0.2
Flask==2.2.2
folium==0.12.1.post1
fonttools==4.38.0
future==0.18.2
geojson==2.5.0
google-api-core==2.10.1
google-api-python-client==1.12.11
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
gpiozero==1.6.2
httplib2==0.20.4
httplib2shim==0.0.3
idna==2.10
ImageHash==4.3.1
importlib-metadata==5.0.0
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
manimce==0.1.1.post2
manimlib==0.2.0
MarkupSafe==2.1.1
matplotlib==3.6.0
mccabe==0.7.0
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
pangocairocffi==0.4.0
pangocffi==0.8.0
picamera2==0.3.3
pidng==4.0.9
piexif==1.1.3
Pillow==9.2.0
platformdirs==2.5.2
progressbar==2.5
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycodestyle==2.9.1
pycparser==2.21
pydub==0.25.1
pyftdi==0.54.0
Pygments==2.7.1
PyGObject==3.38.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyJWT==2.6.0
pylint==2.15.3
pyowm==3.3.0
pyparsing==3.0.9
pyserial==3.5
PySocks==1.7.1
python-apt==2.2.1
python-dateutil==2.8.2
python-prctl==1.7
python3-openid==3.2.0
pytz==2022.5
pyusb==1.2.1
PyWavelets==1.4.1
PyYAML==5.3.1
redis==4.3.4
requests==2.25.1
requests-oauthlib==1.3.1
rgbmatrix==0.0.1
rich==6.2.0
rpi-ws281x==4.3.4
RPi.GPIO==0.7.0
rsa==4.9
scipy==1.9.2
seaborn==0.12.1
simplejpeg==1.6.4
six==1.16.0
social-auth-core==4.3.0
spidev==3.5
spotipy==2.20.0
ssh-import-id==5.10
sysv-ipc==1.1.0
toml==0.10.1
tomli==2.0.1
tomlkit==0.11.5
tqdm==4.64.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.5
v4l2-python3==0.3.1
Werkzeug==2.2.2
wrapt==1.14.1
xdg==5.1.1
zipp==3.8.1
